Decouple core and plugin parsers
