+++ flexget/plugins/parsers/parser_common.py
@@ -25,23 +25,6 @@ def clean_value(name):
     return name
 
 
-class ParseWarning(Warning):
-
-    def __init__(self, parsed, value, **kwargs):
-        self.value = value
-        self.parsed = parsed
-        self.kwargs = kwargs
-
-    def __unicode__(self):
-        return self.value
-
-    def __str__(self):
-        return self.__unicode__().encode('utf-8')
-
-    def __repr__(self):
-        return str('ParseWarning({}, **{})').format(self, repr(self.kwargs))
-
-
 def old_assume_quality(guessed_quality, assumed_quality):
     if assumed_quality:
         if not guessed_quality:
@@ -55,43 +38,6 @@ def old_assume_quality(guessed_quality, assumed_quality):
         if assumed_quality.audio:
             guessed_quality.audio = assumed_quality.audio
     return guessed_quality
-
-
-default_ignore_prefixes = [
-    '(?:\[[^\[\]]*\])',  # ignores group names before the name, eg [foobar] name
-    '(?:HD.720p?:)',
-    '(?:HD.1080p?:)',
-    '(?:HD.2160p?:)'
-]
-
-
-def name_to_re(name, ignore_prefixes=None, parser=None):
-    """Convert 'foo bar' to '^[^...]*foo[^...]*bar[^...]+"""
-    if not ignore_prefixes:
-        ignore_prefixes = default_ignore_prefixes
-    parenthetical = None
-    if name.endswith(')'):
-        p_start = name.rfind('(')
-        if p_start != -1:
-            parenthetical = re.escape(name[p_start + 1:-1])
-            name = name[:p_start - 1]
-    # Blanks are any non word characters except & and _
-    blank = r'(?:[^\w&]|_)'
-    ignore = '(?:' + '|'.join(ignore_prefixes) + ')?'
-    res = re.sub(re.compile(blank + '+', re.UNICODE), ' ', name)
-    res = res.strip()
-    # accept either '&' or 'and'
-    res = re.sub(' (&|and) ', ' (?:and|&) ', res, re.UNICODE)
-    # The replacement has a regex escape in it (\w) which needs to be escaped again in python 3.7+
-    res = re.sub(' +', blank.replace('\\', '\\\\') + '*', res, re.UNICODE)
-    if parenthetical:
-        res += '(?:' + blank + '+' + parenthetical + ')?'
-        # Turn on exact mode for series ending with a parenthetical,
-        # so that 'Show (US)' is not accepted as 'Show (UK)'
-        if parser:
-            parser.strict_name = True
-    res = '^' + ignore + blank + '*' + '(' + res + ')(?:\\b|_)' + blank + '*'
-    return res
 
 
 def remove_dirt(name):
