+++ flexget/plugins/parsers/parser_guessit.py
@@ -1,22 +1,23 @@
 from __future__ import unicode_literals, division, absolute_import
-from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
-from future.utils import native
 
 import logging
 import re
 import sys
 import time
+from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
 
-from guessit.rules import rebulk_builder
+from future.utils import native
 from guessit.api import GuessItApi, GuessitException
+from guessit.rules import rebulk_builder
 from rebulk import Rebulk
 from rebulk.pattern import RePattern
 
 from flexget import plugin
 from flexget.event import event
 from flexget.utils import qualities
+from flexget.utils.parsers.generic import default_ignore_prefixes, name_to_re, ParseWarning
 from flexget.utils.tools import ReList
-from .parser_common import MovieParseResult, SeriesParseResult, default_ignore_prefixes, name_to_re, ParseWarning
+from .parser_common import MovieParseResult, SeriesParseResult
 
 log = logging.getLogger('parser_guessit')
