+++ flexget/plugins/services/pogcal_acquired.py
@@ -1,18 +1,18 @@
 from __future__ import unicode_literals, division, absolute_import
-from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
 
 import logging
 import re
+from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
 from datetime import datetime
 
 from sqlalchemy import Column, Unicode, Integer
 
 from flexget import plugin
+from flexget.db_schema import versioned_base
 from flexget.event import event
 from flexget.utils import requests
+from flexget.utils.parsers.generic import name_to_re
 from flexget.utils.soup import get_soup
-from flexget.utils.titles.series import name_to_re
-from flexget.db_schema import versioned_base
 
 log = logging.getLogger('pogcal_acquired')
 Base = versioned_base('pogcal_acquired', 0)
