+++ flexget/utils/parsers/__init__.py
@@ -1,0 +1,4 @@
+# make importing these a bit less hassle
+from flexget.utils.parsers.series import SeriesParser, ID_TYPES  # noqa pylint: disable=unused-import
+from flexget.utils.parsers.movie import MovieParser  # noqa pylint: disable=unused-import
+from flexget.utils.parsers.parser import TitleParser  # noqa pylint: disable=unused-import
