+++ flexget/utils/parsers/parser.py
