--- flexget/utils/titles/__init__.py
@@ -1,4 +1,0 @@
-# make importing these a bit less hassle
-from flexget.utils.titles.series import SeriesParser, ID_TYPES  # noqa pylint: disable=unused-import
-from flexget.utils.titles.movie import MovieParser  # noqa pylint: disable=unused-import
-from flexget.utils.titles.parser import TitleParser  # noqa pylint: disable=unused-import
